Moved all objects out of farm, and into their own object folder in the gameSetup root
