+++ client/src/components/GameSetup/Objects/Crops/Carrot.js
