+++ client/src/components/GameSetup/Objects/Crops/Potato.js
