+++ client/src/components/GameSetup/Objects/Dirt.js
