+++ client/src/components/GameSetup/Objects/Object.js
