+++ client/src/components/GameSetup/Objects/helpers.js
